requirements.txt
black==20.8b1
flask==1.1.2
Flask-GraphQL==2.0.1
graphene==2.1.8
grpcio==1.33.2
grpcio-tools==1.33.2
